Temp workaround: use floats instead of doubles w/ ComplexDoubleTypes
To be reverted later after merge with master.

+++ imglib2/ops/src/test/java/net/imglib2/ops/example/SpeedCalcs.java
@@ -31,7 +31,7 @@ public class SpeedCalcs {
 		ComplexDoubleType out2 = new ComplexDoubleType();
 		for (double x = -100; x <= 100; x += 0.5) {
 			for (double y = -100; y <= 100; y += 0.3) {
-				ComplexDoubleType input = new ComplexDoubleType(x,y);
+				ComplexDoubleType input = new ComplexDoubleType((float)x,(float)y);
 				cosFunc.compute(input, cos);
 				method1(cos, out1);
 				method2(cos, out2);
